img_processing_dashboard/requirements.txt
# ==============================
# IMAGE PROCESSING DASHBOARD
# Dependencies for Edge Detection App
# ==============================

# Core image processing
opencv-python
numpy
pillow

# Data and math utilities
pandas
scipy

# Visualization / Histogram
matplotlib
plotly

# Web framework (choose one depending on deployment)
streamlit
# flask

# Misc
tqdm

#python -m venv venv
#venv\Scripts\activate
